[Bronze V] Title: 이상한 기호, Time: 168 ms, Memory: 17724 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,10 +5,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int A = scanner.nextInt();
-        int B = scanner.nextInt();
+        long A = scanner.nextLong();
+        long B = scanner.nextLong();
 
-        int result = (A + B) * (A - B);
+        long result = (A + B) * (A - B);
 
         System.out.println(result);
     }
